what about now

+++ pyscripts/test-ar_circuit-calcs.py
@@ -43,7 +43,7 @@ share0prime = np.array(share0prime, dtype=np.uintc)
 share1prime = np.array(share1prime, dtype=np.uintc)
 print("SHARES 0: ", share0)
 for i in range(0, len(x)):
-    print(share0[i]+share1[i])
+    print(np.uintc(share0[i]+share1[i]))
 
 
 output = pffrocd.run_sfe_improved(x, y, y_0=share0, y_1=share1, x_0=share0prime, x_1=share1prime)
